+++ rmpc/src/config/theme/lyrics.rs
@@ -1,18 +1,28 @@
 use serde::{Deserialize, Serialize};
 
+use crate::config::theme::properties::Alignment;
+
 #[derive(Debug, Default, Clone)]
 pub struct LyricsConfig {
     pub timestamp: bool,
+    pub alignment: Alignment,
 }
 
-#[derive(Debug, Serialize, Deserialize, PartialEq, Eq, Default)]
+#[derive(Debug, Serialize, Deserialize, PartialEq, Eq)]
+#[serde(default)]
 pub struct LyricsConfigFile {
-    #[serde(default)]
     pub(super) timestamp: bool,
+    pub(super) alignment: Alignment,
+}
+
+impl Default for LyricsConfigFile {
+    fn default() -> Self {
+        Self { timestamp: false, alignment: Alignment::Center }
+    }
 }
 
 impl From<LyricsConfigFile> for LyricsConfig {
     fn from(value: LyricsConfigFile) -> Self {
-        LyricsConfig { timestamp: value.timestamp }
+        LyricsConfig { timestamp: value.timestamp, alignment: value.alignment }
     }
 }